todo-web-app/backend/pyproject.toml
[project]
name = "todo-backend"
version = "2.2.0"
description = "Phase 2.2 Todo Backend API — JWT Authentication"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.21",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.3.0",
    "alembic>=1.13.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "mcp[cli]>=1.0.0",
    # [Task]: T-3.2.1
    "openai-agents>=0.0.12",    # OpenAI Agents SDK — Groq + MCP tool orchestration
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
